Move the get_resource function into the app module
